+++ src/components/forms/FileUploadBlock.js
@@ -3,6 +3,14 @@ import React, {useState, useEffect} from "react";
 import {Global,css} from "@emotion/react"
 import { Badge } from 'reactstrap';
 import styled from '@emotion/styled'
+import {useMutation, useQueryClient} from 'react-query';
+import {useAppContext} from "../../context";
+import {serviceKyc} from '../../services/kyc.service'
+import {useServiceKyc} from '../../hooks';
+import { useForm } from "react-hook-form";
+import { yupResolver } from '@hookform/resolvers/yup';
+import { kycSchema } from "../../validations";
+import { useRouter } from 'next/router';
 // reactstrap components
 import {
   DropdownToggle,
@@ -15,17 +23,49 @@ import {
 } from "reactstrap";
 import Image from 'next/image';
 
-function FileUploadBlock({setBucket,text,id,idResponse}) {
+function FileUploadBlock({setResponseAlert,setVisibleAlert,text,id,idSend,idResponse}) {
+const {mutateAsync:cpc, isLoading:icpc}  = useServiceKyc();
+const context = useAppContext();
+const router = useRouter();
+const { register, handleSubmit , errors} = useForm({
+  resolver: yupResolver(kycSchema),
+});
+const [datasending, setDatasending] = useState("");
+const [bucket, setBucket] = useState("");
+const [responseToast, setResponseToast] = useState("");
+const [submitting, setSubmitting] = useState(false);
+const onSubmit = async ()  => {
+  const body = {
+    data: {
+      file: datasending,
+      bucket: bucket
+      }
+    }
+      //setSubmitting(isLoading);
+      /*const { kyc_display_input} = hookData;
+      userdata = {kyc_display_input};*/
+      //setDatasending(hookData.kyc_display_input)
+      const res = await cpc({accessToken: context.appState.accessToken ,data:body});
+      if(res.error && !res.success){
+        console.log(res.message)
+        setResponseAlert(res.message);
+        setVisibleAlert(true);
+         } else {
+           setShowToast(true);
+           setResponseAlert(res.message);
+           setVisibleAlert(true);
+       }
+};
   function uploadFiles(files,idResponse){
     const kyc_display = document.getElementById("kyc_display");
-    const kyc_display_input = document.getElementsByName("kyc_display_input")[0];
 	 	let elt = "";
 		let reader = new FileReader();
     	reader.addEventListener("load", function(){
+        setBucket(id);
         requestAnimationFrame(()=>{
-          idResponse.innerHTML = "Un fichier a été selectionné";
+          idResponse.innerHTML = files.name;
           kyc_display.src = this.result;
-          kyc_display_input.value = this.result;
+          setDatasending(this.result);
           kyc_display.style.visibility = "visible";
         });
     	},false);
@@ -33,11 +73,13 @@ function FileUploadBlock({setBucket,text,id,idResponse}) {
     }
   const launchUpload = (id,idResponse) => {
     const response = document.getElementById(idResponse);
+    const send = document.getElementById(idSend);
     const source_file = document.createElement("input");
     source_file.type = "file";
     source_file.multiple = "false";
     source_file.click();
     source_file.addEventListener("change",()=>{
+        setSubmitting(true);
   		uploadFiles(source_file.files[0],response);
   	});
     return source_file;
@@ -65,9 +107,11 @@ function FileUploadBlock({setBucket,text,id,idResponse}) {
                 <Badge style={{float:"right",padding:"0.4em",backgroundColor:"#32DC00"}}color="success"> </Badge></h4>
                     <Row>
                        <Col xs="12"><span onClick={() =>{
-                         setBucket(id);
                          launchUpload(id,idResponse);
                        }} style={{backgroundColor:"#669965",color:"white",cursor:"pointer",padding:"0.4em",borderRadius:"10px",marginRight:"0.8em"}}>Ajouter</span><span id={idResponse}> Aucun fichier choisit</span></Col>
+                    </Row><br/>
+                    <Row>
+                       <Col xs="12"><span style={{display:"visible",backgroundColor:"#c89631",color:"white",cursor:"pointer",padding:"0.4em",borderRadius:"10px",marginRight:"0.8em"}} onClick={() =>{onSubmit()}} id={idSend}>Envoyer</span></Col>
                     </Row>
            </Col>
         </Row>
